requirements.txt
blake2
requests
pycrypto
appdirs
validate_email
